rustc: move Debug impls from ppaux to ty::structural_impls.
