+++ src/librustc/ty/subst.rs
@@ -12,6 +12,7 @@ use smallvec::SmallVec;
 use rustc_macros::HashStable;
 
 use core::intrinsics;
+use std::fmt;
 use std::cmp::Ordering;
 use std::marker::PhantomData;
 use std::mem;
@@ -65,6 +66,16 @@ impl<'tcx> UnpackedKind<'tcx> {
                 NonZeroUsize::new_unchecked(ptr | tag)
             },
             marker: PhantomData
+        }
+    }
+}
+
+impl fmt::Debug for Kind<'tcx> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self.unpack() {
+            UnpackedKind::Lifetime(lt) => lt.fmt(f),
+            UnpackedKind::Type(ty) => ty.fmt(f),
+            UnpackedKind::Const(ct) => ct.fmt(f),
         }
     }
 }
